Implement workspaces
Let's lay out the possible states of SublimeText:

When testing these situations, make sure to delete your session
file at `$sublime_config_dir/Local/Session.sublime-session`.

- Single-file window ("subl /path/to/file"). In this state,
  window.project_file_name() is None and window.project_data() is
  also None. According to the LSP specification, we're allowed to
  report `null` for `workspaces`.
- Window with a bunch of files open ("subl /path/to/dir"). In this
  state, window.project_file_name() is None and window.project_data()
  contains a single folder that points to "/path/to/dir". We can
  report that single folder (as URI) for `workspaces`.
  *This state is the most common situation the user will be in.*
- Window that was opened via a .sublime-project file
  ("subl /path/to/dir/myproject.sublime-project"). In this state,
  window.project_file_name() is an absolute path to the project file.
  window.project_data() is a dictionary of names to (relative)
  folders. We will convert all of the folders to absolute paths,
  make them URIs, and then report that list as the current
  `workspaces`.

There is a complication with the last two points. Namely, when you
have a Sublime window open in the background, and then open a single
file, say "subl /path/to/foo.json", then that single file will become
part of the existing window.

Such a single file may be *outside* the workspaces of the window.

I thought about it for a while, and decided to ignore this
complication. That is, when that single-file starts a new session,
it will still report the workspaces of the window.

1) This makes it transparent to the user what LSP is doing.
2) In my experience, when opening a single file outside the project
   directories, such a file can stand on its own (e.g. a JSON or
   YAML configuration file). A server for such a file should be
   able to handle single files and ignore workspaces.
3) It aligns better with the first situation where a single file
   occupies a single window where window.project_data() is None.
   In that case we'll send `null` for the workspaces so we're
   already expecting the language server to be able to work without
   workspaces.
4) It makes the programming logic much more easier.

There are a few places where we assumed a single root URI. For
example, handling textDocument/references. I have modified these
occurences to take multiple folders into account, by using the
common prefix of all workspaces instead.

Finally, it's also possible that the window folders change when
using the "Quick Switch Project" feature. We should shut down old
sessions and start new sessions in that case. That behavior has
not changed.