Use Sublime's cache path as the working directory for language servers
When the number of workspaces is zero or >1, it is ambiguous what the
working directory of a language server should be. For this reason I think
it's best to use a temp directory.

When running under Sublime Text, the cache path is

    $sublime_cache_path/LSP/$servername

When running under pytest, the cache path is

    $TMP/LSP/$servername